requirements.txt
streamlit
datetime
pillow
google.generativeai